cleanup membercontroller
